+++ client/src/components/App.js
@@ -8,7 +8,7 @@ import LoginPage from './LoginPage'
 import Dashboard from './Dashboard'
 import Statistics from './Statistics'
 import Logs from './Logs'
-import InsertUser from './InsertUser'
+import InsertUser from './InsertUserPage'
 import ModifyUser from './ModifyUser'
 import DeleteUser from './DeleteUser'
 import NotFound from './NotFound'
